Support direflow-config.json file

+++ packages/direflow-scripts/src/config/config-overrides.ts
@@ -14,33 +14,41 @@ import {
   TEntry,
   IPlugin,
 } from '../types/ConfigOverrides';
+import getDireflowConfig from '../helpers/getDireflowConfig';
+import IDireflowConfig from '../types/DireflowConfig';
 
-export = function override(config: TConfig, env: string, options?: IOptions) {
-  const filename = options?.filename || 'direflowBundle.js';
-  const chunkFilename = options?.chunkFilename || 'vendor.js';
-  const react = options?.react;
-  const reactDOM = options?.reactDOM;
+export = function override(config: TConfig, env: string) {
+  const originalEntry = [...(config.entry as string[])];
+  const [pathIndex] =
+    env === 'development' ? originalEntry.splice(1, 1) : originalEntry.splice(0, 1);
 
-  const entries = addEntries(config.entry, env, { react, reactDOM });
+  const direflowConfig = getDireflowConfig(pathIndex);
+
+  const useVendor = !!direflowConfig?.build?.vendor;
+
+  const entries = addEntries(config.entry, pathIndex, env, direflowConfig);
 
   const overridenConfig = {
     ...config,
     entry: entries,
     module: overrideModule(config.module),
-    output: overrideOutput(config.output, env, { filename, chunkFilename }),
-    optimization: overrideOptimization(config.optimization, env),
+    output: overrideOutput(config.output, direflowConfig),
+    optimization: overrideOptimization(config.optimization, env, useVendor),
     resolve: overrideResolve(config.resolve),
-    plugins: overridePlugins(config.plugins, entries, env, { filename, chunkFilename }),
-    externals: overrideExternals(config.externals, env, { react, reactDOM }),
+    plugins: overridePlugins(config.plugins, entries, env, direflowConfig),
+    externals: overrideExternals(config.externals, env, direflowConfig),
   };
 
   return overridenConfig;
 };
 
-function addEntries(entry: TEntry, env: string, { react, reactDOM }: IOptions) {
+function addEntries(entry: TEntry, pathIndex: string, env: string, config?: IDireflowConfig) {
   const originalEntry = [...(entry as string[])];
 
-  const [pathIndex] = env === 'development' ? originalEntry.splice(1, 1) : originalEntry.splice(0, 1);
+  const react = config?.modules?.react;
+  const reactDOM = config?.modules?.reactDOM;
+  const useSplit = !!config?.build?.split;
+
   const resolvedEntries = entryResolver(pathIndex, { react, reactDOM });
 
   const newEntry: { [key: string]: string } = { main: pathIndex };
@@ -61,7 +69,7 @@ function addEntries(entry: TEntry, env: string, { react, reactDOM }: IOptions) {
     return [...flatList, resolve(__dirname, '../template-scripts/welcome.js')];
   }
 
-  if (hasOptions('split', env)) {
+  if (useSplit) {
     return newEntry;
   }
 
@@ -88,12 +96,12 @@ function overrideModule(module: IModule) {
   return module;
 }
 
-function overrideOutput(
-  output: IOptions,
-  env: string,
-  { filename, chunkFilename }: IOptions,
-) {
-  const outputFilename = hasOptions('split', env) ? '[name].js' : filename;
+function overrideOutput(output: IOptions, config?: IDireflowConfig) {
+  const useSplit = config?.build?.split;
+  const filename = config?.build?.filename || 'direflowBundle.js';
+  const chunkFilename = config?.build?.chunkFilename || 'vendor.js';
+
+  const outputFilename = useSplit ? '[name].js' : filename;
 
   return {
     ...output,
@@ -102,7 +110,7 @@ function overrideOutput(
   };
 }
 
-function overrideOptimization(optimization: IOptimization, env: string) {
+function overrideOptimization(optimization: IOptimization, env: string, useVendor: boolean) {
   optimization.minimizer[0].options.sourceMap = env === 'development';
 
   const vendorSplitChunks = {
@@ -118,17 +126,17 @@ function overrideOptimization(optimization: IOptimization, env: string) {
 
   return {
     ...optimization,
-    splitChunks: hasOptions('vendor', env) ? vendorSplitChunks : false,
+    splitChunks: useVendor ? vendorSplitChunks : false,
     runtimeChunk: false,
   };
 }
 
-function overridePlugins(plugins: IPlugin[], entry: TEntry, env: string, options: IOptions) {
+function overridePlugins(plugins: IPlugin[], entry: TEntry, env: string, config?: IDireflowConfig) {
   plugins[0].options.inject = 'head';
 
   plugins.push(
     new EventHooksPlugin({
-      done: new PromiseTask(() => copyBundleScript(env, entry, options)),
+      done: new PromiseTask(() => copyBundleScript(env, entry, config)),
     }),
   );
 
@@ -158,30 +166,46 @@ function overrideResolve(currentResolve: IResolve) {
   return currentResolve;
 }
 
-function overrideExternals(externals: { [key: string]: any }, env: string, { react, reactDOM }: IOptions) {
+function overrideExternals(
+  externals: { [key: string]: any },
+  env: string,
+  config?: IDireflowConfig,
+) {
   if (env === 'development') {
     return externals;
   }
 
   const extraExternals: any = { ...externals };
+  const react = config?.modules?.react;
+  const reactDOM = config?.modules?.reactDOM;
 
-  if (react !== false) {
+  if (react) {
     extraExternals.react = 'React';
   }
 
-  if (reactDOM !== false) {
+  if (reactDOM) {
     extraExternals['react-dom'] = 'ReactDOM';
   }
 
   return extraExternals;
 }
 
-async function copyBundleScript(env: string, entry: TEntry, { filename, chunkFilename }: IOptions) {
+async function copyBundleScript(env: string, entry: TEntry, config?: IDireflowConfig) {
   if (env !== 'production') {
     return;
   }
 
   if (!fs.existsSync('build')) {
+    return;
+  }
+
+  const filename = config?.build?.filename || 'direflowBundle.js';
+  const chunkFilename = config?.build?.chunkFilename || 'vendor.js';
+  const emitAll = config?.build?.emitAll;
+  const emitSourceMaps = config?.build?.emitSourceMap;
+  const emitIndexHTML = config?.build?.emitIndexHTML;
+
+  if (emitAll) {
     return;
   }
 
@@ -198,22 +222,14 @@ async function copyBundleScript(env: string, entry: TEntry, { filename, chunkFil
       return;
     }
 
+    if (emitSourceMaps && file.endsWith('.map')) {
+      return;
+    }
+
+    if (emitIndexHTML && file.endsWith('.html')) {
+      return;
+    }
+
     rimraf.sync(`build/${file}`);
   });
-}
-
-function hasOptions(flag: string, env: string) {
-  if (env !== 'production') {
-    return false;
-  }
-
-  if (process.argv.length < 3) {
-    return false;
-  }
-
-  if (!process.argv.some((arg: string) => arg === `--${flag}` || arg === `-${flag[0]}`)) {
-    return false;
-  }
-
-  return true;
 }
